+++ rocAL/include/augmentations/filter_augmentations/node_erase.h
@@ -1,5 +1,5 @@
 /*
-Copyright (c) 2019 - 2025 Advanced Micro Devices, Inc. All rights reserved.
+Copyright (c) 2025 Advanced Micro Devices, Inc. All rights reserved.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -21,28 +21,39 @@ THE SOFTWARE.
 */
 
 #pragma once
+#include <list>
+#include <VX/vx.h>
+
 #include "pipeline/graph.h"
 #include "pipeline/node.h"
-#include "parameters/parameter_factory.h"
 #include "parameters/parameter_vx.h"
 
-class BrightnessNode : public Node {
+class EraseNode : public Node {
    public:
-    BrightnessNode(const std::vector<Tensor *> &inputs, const std::vector<Tensor *> &outputs);
-    BrightnessNode() = delete;
+    EraseNode(const std::vector<Tensor *> &inputs, const std::vector<Tensor *> &outputs);
+    EraseNode() = delete;
+    ~EraseNode();
 
-    void init(float alpha, float beta);
-    void init(FloatParam *alpha_param, FloatParam *beta_param);
-    void initialize_args(const ArgumentSet& arguments) override;
-    std::string node_name() const override { return "BrightnessNode"; }
+    void init(Tensor *anchor_box_info, Tensor *colors, int num_boxes_fixed);
+    void init(Tensor *anchor_box_info, Tensor *colors, IntParam *num_boxes_param);
+    void init(std::vector<float> anchor,
+              std::vector<float> shape,
+              std::vector<unsigned> num_boxes,
+              std::vector<float> fill_value);
 
    protected:
     void create_node() override;
     void update_node() override;
 
    private:
-    ParameterVX<float> _alpha;
-    ParameterVX<float> _beta;
-    constexpr static float ALPHA_RANGE[2] = {0.1, 1.95};
-    constexpr static float BETA_RANGE[2] = {0, 25};
+    std::vector<int> _anchor_vec;
+    std::vector<float> _colors_vec, _fill_values_vec, _fill_values;
+    std::vector<unsigned> _num_boxes_vec;
+    vx_tensor _anchor_vx = nullptr;
+    vx_tensor _colors_vx = nullptr;
+    vx_tensor  _num_boxes_vx = nullptr;
+    void* _anchor_ptr = nullptr;
+    void* _color_ptr = nullptr;
+    void* _num_box_ptr = nullptr;
+    unsigned _total_boxes = 0;
 };